[Chore] Swift 4.2 Update Update required files to support swift 4.2 Yoshi version updated to 3.0.0-beta.2

+++ Yoshi/Yoshi/YoshiConfigurationManager.swift
@@ -54,7 +54,7 @@ internal final class YoshiConfigurationManager {
         }
 
         let window = UIWindow(frame: UIScreen.main.bounds)
-        window.windowLevel = UIWindowLevelNormal
+        window.windowLevel = CGFloat(UIWindowLevelNormal)
 
         // Use a dummy view controller with clear background.
         // This way, we can make the actual view controller we want to present a form sheet on the iPad.
